requirements.txt
streamlit
mcp
python-dotenv
aiohttp